+++ src/Bootstrap/ReplaceTemplateTags.php
@@ -23,6 +23,8 @@ class ReplaceTemplateTags
             $exp->registerTag('channel.categories', \Expressionengine\Coilpack\View\Tags\Channel\Categories::class);
             $exp->registerTag('channel.entries', \Expressionengine\Coilpack\View\Tags\Channel\Entries::class);
             $exp->registerTag('comment.entries', \Expressionengine\Coilpack\View\Tags\Comment\Entries::class);
+            $exp->registerTag('comment.form', \Expressionengine\Coilpack\View\Tags\Comment\Form::class);
+            $exp->registerTag('comment.preview', \Expressionengine\Coilpack\View\Tags\Comment\Preview::class);
             $exp->registerTag('structure.entries', \Expressionengine\Coilpack\View\Tags\Structure\Entries::class);
             $exp->registerTag('structure.nav', \Expressionengine\Coilpack\View\Tags\Structure\Nav::class);
             $exp->registerTag('email.contact_form', \Expressionengine\Coilpack\View\Tags\Email\ContactForm::class);